[wmco] Delete existing service monitor object on WMCO restart
This commit checks for an existing service monitor object on a WMCO restart.
If a service monitor object does exist, it is deleted and a new one is created.
We are deleting to ensure that the SM always exists with the correct spec. Otherwise,
metrics may exhibit unexpected behavior if created by a previous version of WMCO.

(cherry picked from commit 2209daf0d0c6d5ebe7a0a986c509854539355780)